requirements.txt
asgiref==3.3.4

Django==3.0
djangorestframework==3.12.4
django-allauth==0.44.0
django-rest-auth==0.9.5
django-model-utils>=2.0.3
dj-database-url==0.5.0
psycopg2>=2.8,<2.9
pytz==2021.1
sqlparse==0.4.1

# heroku config
django-heroku==0.3.1
gunicorn==20.1.0 
whitenoise==5.2.0

# For adding JWT support
djangorestframework-jwt==1.11.0
PyJWT==1.7.1

# json fields
jsonfield>=1.0.3

pytest-django==3.9.0
six==1.16.0

# For enabling cors headers
django-cors-headers==3.7.0 

# For adding images to models
pillow==8.3.2

# For auto generating one to one fields
django-annoying==0.10.6

# Add Firebase admin sdk
firebase-admin==5.0.0